+++ android/feature/home/src/main/res/layout/fragment_home.xml
@@ -17,7 +17,6 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:fitsSystemWindows="true"
             android:paddingBottom="32dp">
 
             <com.google.android.material.appbar.MaterialToolbar
